delete offers, reservation

+++ api/Repository/ReservationRepository.cs
@@ -24,10 +24,11 @@ namespace api.Repository
             return reservationModel;
         }
 
-        public async Task<Reservation> DeleteReservationAsync(int id)
+        public async Task<Reservation> DeleteReservationAsync(int id,int userId)
         {
             var reservation = await _context.Reservation.FirstOrDefaultAsync(r=>r.Id==id);
             if(reservation==null)return null;
+            if(userId!=reservation.userId) return null;
             _context.Reservation.Remove(reservation);
             await _context.SaveChangesAsync();
             return reservation;
